tether and weapon spawns
Added tethers to the map that are used for spawning tethers. I do not know how enemies will spawn yet, but we could use the tethers to spawn enemies in the future if needed. Right now a box just spawns as a weapon and nothing happens when they picked up yet.