New levels, and comp change
New levels have been added and competitve now rolls a random build index within the range for the levels. If the random.range rolls the same level the player is currently on then it rerolls until it finds a value that is not the same as the current level index.